+++ capstone-rs/src/test.rs
@@ -11,7 +11,7 @@ use alloc::vec::Vec;
 #[cfg(feature = "full")]
 use {alloc::string::String, std::collections::HashSet};
 
-use capstone_sys::{arm64_op_sme_index, cs_group_type};
+use capstone_sys::cs_group_type;
 use libc::c_uint;
 
 use super::arch::*;
@@ -1125,7 +1125,7 @@ fn test_arch_arm64_detail() {
     //use crate::arch::arm64::Arm64Sysreg::*;
     use crate::arch::arm64::Arm64Vas::*;
     use crate::arch::arm64::*;
-    use capstone_sys::arm64_op_mem;
+    use capstone_sys::{arm64_op_mem, arm64_op_sme_index};
 
     let s0 = Arm64Operand {
         op_type: Reg(RegId(ARM64_REG_S0 as RegIdInt)),
@@ -2189,6 +2189,140 @@ fn test_arch_mips_detail() {
 }
 
 #[test]
+fn test_arch_mos65xx() {
+    test_arch_mode_endian_insns(
+        &mut Capstone::new()
+            .mos65xx()
+            .mode(mos65xx::ArchMode::Mos65xx6502)
+            .build()
+            .unwrap(),
+        Arch::MOS65XX,
+        Mode::Mos65xx6502,
+        None,
+        &[],
+        &[("lda", b"\xa1\xa2"), ("ora", b"\x0d\x34\x12")],
+    );
+
+    test_arch_mode_endian_insns(
+        &mut Capstone::new()
+            .mos65xx()
+            .mode(mos65xx::ArchMode::Mos65xx65c02)
+            .build()
+            .unwrap(),
+        Arch::MOS65XX,
+        Mode::Mos65xx65c02,
+        None,
+        &[],
+        &[
+            ("inc", b"\x1a"),
+            ("dec", b"\x3a"),
+            ("nop", b"\x02\x12"),
+            ("nop", b"\x03"),
+            ("nop", b"\x5c\x34\x12"),
+        ],
+    );
+
+    test_arch_mode_endian_insns(
+        &mut Capstone::new()
+            .mos65xx()
+            .mode(mos65xx::ArchMode::Mos65xxW65c02)
+            .build()
+            .unwrap(),
+        Arch::MOS65XX,
+        Mode::Mos65xxW65c02,
+        None,
+        &[],
+        &[("rmb0", b"\x07\x12"), ("rmb2", b"\x27\x12")],
+    );
+
+    test_arch_mode_endian_insns(
+        &mut Capstone::new()
+            .mos65xx()
+            .mode(mos65xx::ArchMode::Mos65xx65816LongMx)
+            .build()
+            .unwrap(),
+        Arch::MOS65XX,
+        Mode::Mos65xx65816LongMx,
+        None,
+        &[],
+        &[("lda", b"\xa9\x34\x12"), ("mvp", b"\x44\x34\x12")],
+    );
+}
+
+#[test]
+fn test_arch_mos65xx_detail() {
+    use crate::arch::mos65xx::Mos65xxOperand::*;
+    use capstone_sys::mos65xx_reg::*;
+
+    test_arch_mode_endian_insns_detail(
+        &mut Capstone::new()
+            .mos65xx()
+            .mode(mos65xx::ArchMode::Mos65xx6502)
+            .build()
+            .unwrap(),
+        Arch::MOS65XX,
+        Mode::Mos65xx6502,
+        None,
+        &[],
+        &[
+            DII::new("lda", b"\xa1\xa2", &[Mem(0xa2)]),
+            DII::new("ora", b"\x0d\x34\x12", &[Mem(0x1234)]),
+        ],
+    );
+
+    test_arch_mode_endian_insns_detail(
+        &mut Capstone::new()
+            .mos65xx()
+            .mode(mos65xx::ArchMode::Mos65xx65c02)
+            .build()
+            .unwrap(),
+        Arch::MOS65XX,
+        Mode::Mos65xx65c02,
+        None,
+        &[],
+        &[
+            DII::new("inc", b"\x1a", &[Reg(RegId(MOS65XX_REG_ACC as RegIdInt))]),
+            DII::new("dec", b"\x3a", &[Reg(RegId(MOS65XX_REG_ACC as RegIdInt))]),
+            DII::new("nop", b"\x02\x12", &[]),
+            DII::new("nop", b"\x03", &[]),
+            DII::new("nop", b"\x5c\x34\x12", &[]),
+        ],
+    );
+
+    test_arch_mode_endian_insns_detail(
+        &mut Capstone::new()
+            .mos65xx()
+            .mode(mos65xx::ArchMode::Mos65xxW65c02)
+            .build()
+            .unwrap(),
+        Arch::MOS65XX,
+        Mode::Mos65xxW65c02,
+        None,
+        &[],
+        &[
+            DII::new("rmb0", b"\x07\x12", &[Mem(0x12)]),
+            DII::new("rmb2", b"\x27\x12", &[Mem(0x12)]),
+        ],
+    );
+
+    test_arch_mode_endian_insns_detail(
+        &mut Capstone::new()
+            .mos65xx()
+            .mode(mos65xx::ArchMode::Mos65xx65816LongMx)
+            .build()
+            .unwrap(),
+        Arch::MOS65XX,
+        Mode::Mos65xx65816LongMx,
+        None,
+        &[],
+        &[
+            DII::new("lda", b"\xa9\x34\x12", &[Imm(0x1234)]),
+            DII::new("mvp", b"\x44\x34\x12", &[Mem(0x12), Mem(0x34)]),
+        ],
+    );
+}
+
+#[test]
 fn test_arch_ppc() {
     test_arch_mode_endian_insns(
         &mut Capstone::new()
@@ -2329,6 +2463,84 @@ fn test_arch_ppc_detail() {
                 "bgelrl-",
                 b"\x4c\xc8\x00\x21",
                 &[Reg(RegId(PPC_REG_CR2 as RegIdInt))],
+            ),
+        ],
+    );
+}
+
+#[test]
+fn test_arch_sh() {
+    test_arch_mode_endian_insns(
+        &mut Capstone::new()
+            .sh()
+            .mode(sh::ArchMode::Sh4a)
+            .build()
+            .unwrap(),
+        Arch::SH,
+        Mode::Sh4a,
+        None,
+        &[],
+        &[
+            ("add", b"\x0c\x31"),
+            ("mov.b", b"\x10\x20"),
+            ("mov.l", b"\x22\x21"),
+        ],
+    );
+}
+
+#[test]
+fn test_arch_sh_detail() {
+    use crate::arch::sh::ShOpMem;
+    use crate::arch::sh::ShOperand;
+    use capstone_sys::sh_op_mem;
+    use capstone_sys::sh_reg;
+    use capstone_sys::sh_reg::*;
+
+    test_arch_mode_endian_insns_detail(
+        &mut Capstone::new()
+            .sh()
+            .mode(sh::ArchMode::Sh4a)
+            .build()
+            .unwrap(),
+        Arch::SH,
+        Mode::Sh4a,
+        None,
+        &[],
+        &[
+            // add r0, r1
+            DII::new(
+                "add",
+                b"\x0c\x31",
+                &[
+                    ShOperand::Reg(RegId(SH_REG_R0 as RegIdInt)),
+                    ShOperand::Reg(RegId(SH_REG_R1 as RegIdInt)),
+                ],
+            ),
+            // mov.b r1,@r0
+            DII::new(
+                "mov.b",
+                b"\x10\x20",
+                &[
+                    ShOperand::Reg(RegId(SH_REG_R1 as RegIdInt)),
+                    ShOperand::Mem(ShOpMem(sh_op_mem {
+                        address: capstone_sys::sh_op_mem_type::SH_OP_MEM_REG_IND,
+                        reg: SH_REG_R0 as sh_reg::Type,
+                        disp: 0,
+                    })),
+                ],
+            ),
+            // mov.l @r3+,r4
+            DII::new(
+                "mov.l",
+                b"\x36\x64",
+                &[
+                    ShOperand::Mem(ShOpMem(sh_op_mem {
+                        address: capstone_sys::sh_op_mem_type::SH_OP_MEM_REG_POST,
+                        reg: SH_REG_R3 as sh_reg::Type,
+                        disp: 0,
+                    })),
+                    ShOperand::Reg(RegId(SH_REG_R4 as RegIdInt)),
+                ],
             ),
         ],
     );
@@ -3726,7 +3938,7 @@ fn test_arch_tricore() {
             .mode(tricore::ArchMode::TriCore162)
             .build()
             .unwrap(),
-        Arch::TriCore,
+        Arch::TRICORE,
         Mode::TriCore162,
         None,
         &[],
@@ -3747,7 +3959,7 @@ fn test_arch_tricore_detail() {
             .mode(tricore::ArchMode::TriCore162)
             .build()
             .unwrap(),
-        Arch::TriCore,
+        Arch::TRICORE,
         Mode::TriCore162,
         None,
         &[],