+++ capstone-sys/build.rs
@@ -121,12 +121,15 @@ fn build_capstone_cc() {
         .define("CAPSTONE_USE_SYS_DYN_MEM", None)
         .define("CAPSTONE_HAS_ARM", None)
         .define("CAPSTONE_HAS_ARM64", None)
+        .define("CAPSTONE_HAS_BPF", None)
         .define("CAPSTONE_HAS_EVM", None)
         .define("CAPSTONE_HAS_M680X", None)
         .define("CAPSTONE_HAS_M68K", None)
         .define("CAPSTONE_HAS_MIPS", None)
+        .define("CAPSTONE_HAS_MOS65XX", None)
         .define("CAPSTONE_HAS_POWERPC", None)
         .define("CAPSTONE_HAS_RISCV", None)
+        .define("CAPSTONE_HAS_SH", None)
         .define("CAPSTONE_HAS_SPARC", None)
         .define("CAPSTONE_HAS_SYSZ", None)
         .define("CAPSTONE_HAS_TMS320C64X", None)
@@ -134,7 +137,6 @@ fn build_capstone_cc() {
         .define("CAPSTONE_HAS_WASM", None)
         .define("CAPSTONE_HAS_X86", None)
         .define("CAPSTONE_HAS_XCORE", None)
-        .define("CAPSTONE_HAS_BPF", None)
         // No need to display any warnings from the C library
         .flag_if_supported("-w")
         .static_crt(use_static_crt);