+++ capstone-sys/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [UNRELEASED] - YYYY-MM-DD
 ### Added
 - Support for TriCore arch
+- Support for MOS65XX arch
+- Support for SH arch
 
 ### Changed
 - Bump bundled capstone to 5.0.6